requirements.txt
asgiref==3.8.1
asttokens==3.0.0
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.2.1
colorama==0.4.6
comm==0.2.2
Counter==1.0.0
debugpy==1.8.14
decorator==5.2.1
Django==5.2.4
djangorestframework==3.16.0
djangorestframework-simplejwt
dnspython==2.6.1
drf-yasg==1.21.10
et_xmlfile==2.0.0
executing==2.2.0
Flask==3.1.1
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
ipykernel==6.29.5
ipynbname==2024.1.0.0
ipython==9.3.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.8.1
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mysql-connector-python==9.1.0
nest-asyncio==1.6.0
numpy==2.3.0
openpyxl==3.1.5
packaging==25.0
pandas==2.3.0
parso==0.8.4
platformdirs==4.3.8
pluggy==1.6.0
prompt_toolkit==3.0.51
psutil==7.0.0
pure_eval==0.2.3
Pygments==2.19.1
pymongo==4.10.1
pytest==8.4.1
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==310
PyYAML==6.0.2
pyzmq==27.0.0
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
tornado==6.5.1
traitlets==5.14.3
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
wcwidth==0.2.13
Werkzeug==3.1.3
django-cors-headers=4.9.0
pillow=11.3.0